running theory is that authorization will work manually but still figuring out tests

+++ app/controllers/api/v1/jira_sessions_controller.rb
@@ -6,12 +6,13 @@ module Api
     class JiraSessionsController < ApplicationController
       include FormAuth
 
+      before_action :handle_auth, only: %i[authorize]
+      before_action :handle_login, only: :callback
       before_action :form_auth_token
-      before_action :fetch_jira_client, only: %i[new authorize]
       before_action :fetch_session, only: :destroy
 
       def new
-        callback_url = 'http://localhost:3000'
+        callback_url = 'http://localhost:3000/callback'
         request_token = @jira_client.request_token(oauth_callback: callback_url)
         session[:request_token] = request_token.token
         session[:request_secret] = request_token.secret
@@ -19,20 +20,22 @@ module Api
         redirect_to request_token.authorize_url
       end
 
-      def authorize
-        jira_service = JiraService.new(@jira_client)
-        jira_auth = jira_service.create_session(session[:request_token], session[:request_secret])
-
-        session[:jira_auth] = jira_auth
-
-        session.delete(:request_token)
-        session.delete(:request_secret)
-
-        redirect_to projects_path
-      end
+      def authorize;end
 
       def callback
-        redirect_to jira_issues_path
+        if @jira_client
+          jira_service = JiraService.new(@jira_client)
+          jira_auth = jira_service.create_session(session[:request_token], session[:request_secret])
+
+          session[:jira_auth] = jira_auth
+
+          session.delete(:request_token)
+          session.delete(:request_secret)
+
+          render json: { success: 'authorization successful' }
+        else
+          render json: { error: 'Jira client not created' }, status: 500
+        end
       end
 
       def destroy
@@ -40,6 +43,8 @@ module Api
         jira_service.delete_session(@session)
 
         redirect_to skills_path
+
+        return
       end
     end
   end
